centralize cli utils
